@@ -1,10 +1,9 @@
-﻿namespace Lombiq.BaseTheme.Constants
-{
-    public static class ResourceNames
-    {
-        private const string Prefix = FeatureIds.Area + ".";
+﻿namespace Lombiq.BaseTheme.Constants;
 
-        public const string Site = Prefix + nameof(Site);
-        public const string Helpers = Prefix + nameof(Helpers);
-    }
+public static class ResourceNames
+{
+    private const string Prefix = FeatureIds.Area + ".";
+
+    public const string Site = Prefix + nameof(Site);
+    public const string Helpers = Prefix + nameof(Helpers);
 }
