@@ -1,18 +1,17 @@
-﻿namespace Lombiq.BaseTheme.Constants
+﻿namespace Lombiq.BaseTheme.Constants;
+
+public static class ZoneNames
 {
-    public static class ZoneNames
-    {
-        public const string Header = nameof(Header);
-        public const string Navigation = nameof(Navigation);
-        public const string BeforeMain = nameof(BeforeMain);
-        public const string Featured = nameof(Featured);
-        public const string AsideFirst = nameof(AsideFirst);
-        public const string Messages = nameof(Messages);
-        public const string BeforeContent = nameof(BeforeContent);
-        public const string Content = nameof(Content);
-        public const string AfterContent = nameof(AfterContent);
-        public const string AsideSecond = nameof(AsideSecond);
-        public const string AfterMain = nameof(AfterMain);
-        public const string Footer = nameof(Footer);
-    }
+    public const string Header = nameof(Header);
+    public const string Navigation = nameof(Navigation);
+    public const string BeforeMain = nameof(BeforeMain);
+    public const string Featured = nameof(Featured);
+    public const string AsideFirst = nameof(AsideFirst);
+    public const string Messages = nameof(Messages);
+    public const string BeforeContent = nameof(BeforeContent);
+    public const string Content = nameof(Content);
+    public const string AfterContent = nameof(AfterContent);
+    public const string AsideSecond = nameof(AsideSecond);
+    public const string AfterMain = nameof(AfterMain);
+    public const string Footer = nameof(Footer);
 }
