@@ -2,34 +2,33 @@ using Lombiq.BaseTheme.Constants;
 using Microsoft.Extensions.Options;
 using OrchardCore.ResourceManagement;
 
-namespace Lombiq.BaseTheme
+namespace Lombiq.BaseTheme;
+
+public class ResourceManagementOptionsConfiguration : IConfigureOptions<ResourceManagementOptions>
 {
-    public class ResourceManagementOptionsConfiguration : IConfigureOptions<ResourceManagementOptions>
+    private const string WwwRoot = "~/" + FeatureIds.Area + "/";
+    private const string Css = WwwRoot + "css/";
+    private const string Js = WwwRoot + "js/";
+    private const string Vendors = WwwRoot + "vendors/";
+
+    private static readonly ResourceManifest _manifest = new();
+
+    static ResourceManagementOptionsConfiguration()
     {
-        private const string WwwRoot = "~/" + FeatureIds.Area + "/";
-        private const string Css = WwwRoot + "css/";
-        private const string Js = WwwRoot + "js/";
-        private const string Vendors = WwwRoot + "vendors/";
+        _manifest.DefineResource("$" + nameof(FeatureIds.Area), FeatureIds.Area);
 
-        private static readonly ResourceManifest _manifest = new();
+        _manifest
+            .DefineStyle(ResourceNames.Site)
+            .SetUrl(Css + "site.min.css", Css + "site.css");
 
-        static ResourceManagementOptionsConfiguration()
-        {
-            _manifest.DefineResource("$" + nameof(FeatureIds.Area), FeatureIds.Area);
+        _manifest
+            .DefineScript(ResourceNames.Helpers)
+            .SetUrl(Js + "helpers.js");
 
-            _manifest
-                .DefineStyle(ResourceNames.Site)
-                .SetUrl(Css + "site.min.css", Css + "site.css");
-
-            _manifest
-                .DefineScript(ResourceNames.Helpers)
-                .SetUrl(Js + "helpers.js");
-
-            _manifest
-                .DefineScript("bootstrap")
-                    .SetUrl(Vendors + "bootstrap/js/bootstrap.min.js", Vendors + "bootstrap/js/bootstrap.js");
-        }
-
-        public void Configure(ResourceManagementOptions options) => options.ResourceManifests.Add(_manifest);
+        _manifest
+            .DefineScript("bootstrap")
+            .SetUrl(Vendors + "bootstrap/js/bootstrap.min.js", Vendors + "bootstrap/js/bootstrap.js");
     }
+
+    public void Configure(ResourceManagementOptions options) => options.ResourceManifests.Add(_manifest);
 }
